backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.1.1
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.38
threadpoolctl==3.5.0
tifffile==2025.1.10
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
